feat: update memtable api
