+++ src/service/traces/flusher.rs
@@ -1,18 +1,14 @@
 use std::{
     io::{Error, ErrorKind},
-    sync::Arc,
     time::Duration,
 };
 
-use actix_web::web;
 use config::ider;
 use crossbeam_channel::{Receiver as CrossbeamReceiver, Sender as CrossbeamSender};
 use hashbrown::HashMap;
 use infra::errors::BufferWriteError;
-use ingester::Writer;
 use log::{info, warn};
 use once_cell::sync::Lazy;
-use opentelemetry_proto::tonic::collector::trace::v1::ExportTraceServiceRequest;
 use parking_lot::Mutex;
 use tokio::{
     runtime::Runtime,
@@ -22,13 +18,58 @@ use tokio::{
 };
 
 use crate::{
-    common::meta::{stream::SchemaRecords, traces::ExportTraceServiceResponse},
-    service::traces::{flusher, handle_trace_json_request, handle_trace_request},
+    common::meta::{
+        stream::SchemaRecords,
+        traces::{ExportTracePartialSuccess, ExportTraceServiceResponse},
+    },
+    service::{
+        ingestion::TriggerAlertData,
+        metadata::MetadataItem,
+        traces::{flusher, handle_trace_request},
+    },
 };
 
 const BUFFER_FLUSH_INTERVAL: Duration = Duration::from_millis(10);
 // The maximum number of buffered writes that can be queued up before backpressure is applied
-pub const BUFFER_CHANNEL_LIMIT: usize = 10_000;
+const BUFFER_CHANNEL_LIMIT: usize = 10_000;
+
+#[derive(Debug, Clone)]
+pub enum ExportRequest {
+    // Entry((org_id, thread_id, stream_name, entry, trigger, distinct_value, trace_index,
+    // partical_success))
+    TraceEntry(ExportRequestInnerEntry),
+}
+#[derive(Debug)]
+pub struct BufferedWrite {
+    pub request: ExportRequest,
+    pub response_tx: oneshot::Sender<BufferedWriteResult>,
+}
+#[derive(Debug, Clone)]
+pub enum BufferedWriteResult {
+    Success(ExportTraceServiceResponse),
+    Error(BufferWriteError),
+}
+
+type RequestOps = HashMap<String, ExportRequest>;
+pub type ExportRequestInnerEntry = (
+    String,
+    usize,
+    String,
+    std::collections::HashMap<String, SchemaRecords>,
+    Option<TriggerAlertData>,
+    Vec<MetadataItem>,
+    Vec<MetadataItem>,
+    ExportTracePartialSuccess,
+    bool,
+);
+pub struct TraceServiceResponse {
+    // org_id: String,
+    // stream_name: String,
+    response: Result<ExportTraceServiceResponse, BufferWriteError>,
+}
+type NotifyResult = HashMap<String, TraceServiceResponse>;
+type IoFlushNotifyResult = Result<NotifyResult, BufferWriteError>;
+pub static SYNC_RT: Lazy<Runtime> = Lazy::new(|| Runtime::new().unwrap());
 
 #[derive(Debug)]
 pub struct WriteBufferFlusher {
@@ -98,7 +139,7 @@ impl WriteBufferFlusher {
                     BufferedWriteResult::Error(e) => {
                         log::error!("flusher inside write resp error {e}");
                         Err(Error::new(ErrorKind::Other, e))
-                    },
+                    }
                 }
             }
             Err(e) => {
@@ -112,44 +153,6 @@ impl WriteBufferFlusher {
         self.shutdown_tx.clone()
     }
 }
-#[derive(Debug)]
-pub enum ExportRequest {
-    GrpcExportTraceServiceRequest(
-        (
-            String,
-            usize,
-            tonic::Request<ExportTraceServiceRequest>,
-            Option<String>,
-        ),
-    ),
-    HttpJsonExportTraceServiceRequest((String, usize, web::Bytes, Option<String>)),
-}
-#[derive(Debug)]
-pub struct BufferedWrite {
-    pub request: ExportRequest,
-    pub response_tx: oneshot::Sender<BufferedWriteResult>,
-}
-#[derive(Debug, Clone)]
-pub enum BufferedWriteResult {
-    Success(ExportTraceServiceResponse),
-    Error(BufferWriteError),
-}
-
-type RequestOps = HashMap<String, ExportRequest>;
-type NotifyResult = HashMap<
-    String,
-    Result<
-        (
-            ExportTraceServiceResponse,
-            Arc<Writer>,
-            std::collections::HashMap<String, SchemaRecords>,
-            String,
-        ),
-        BufferWriteError,
-    >,
->;
-type IoFlushNotifyResult = Result<NotifyResult, BufferWriteError>;
-pub static SYNC_RT: Lazy<Runtime> = Lazy::new(|| Runtime::new().unwrap());
 
 pub fn run_trace_io_flush(
     io_flush_rx: CrossbeamReceiver<RequestOps>,
@@ -165,37 +168,32 @@ pub fn run_trace_io_flush(
             }
         };
 
-        let mut res = HashMap::new();
+        let mut res: NotifyResult = HashMap::new();
         // write the ops to the segment files, or return on first error
         for (session_id, request) in request {
-            let resp = match request {
-                ExportRequest::GrpcExportTraceServiceRequest(r) => SYNC_RT.block_on(async {
-                    handle_trace_request(
-                        r.0.as_str(),
-                        r.1,
-                        r.2.into_inner(),
-                        true,
-                        r.3.as_deref(),
-                        session_id.as_str(),
-                    )
-                    .await
-                }),
-                ExportRequest::HttpJsonExportTraceServiceRequest(r) => {
-                    let in_stream_name = r.3.unwrap_or("".to_string());
-                    SYNC_RT.block_on(async {
-                        handle_trace_json_request(
-                            r.0.as_str(),
-                            r.1,
-                            r.2,
-                            Some(in_stream_name.as_str()),
-                        )
-                        .await
-                    })
-                }
+            let mut sr = TraceServiceResponse {
+                response: Ok(Default::default()),
+            };
+
+            sr.response = match request {
+                // Entry((org_id, thread_id, stream_name, entry, trigger, distinct_value,
+                // trace_index, partical_success, is_grpc))
+                ExportRequest::TraceEntry(entry) => SYNC_RT.block_on(handle_trace_request(
+                    entry.0.as_str(),
+                    entry.1,
+                    entry.2.as_str(),
+                    entry.3,
+                    entry.4,
+                    entry.5,
+                    entry.6,
+                    entry.7,
+                    session_id.as_str(),
+                    entry.8,
+                )),
             };
 
             // Httpresponse may be partial_success, it must handle every response result
-            res.insert(session_id, resp);
+            res.insert(session_id, sr);
         }
 
         io_flush_notify_tx
@@ -218,51 +216,34 @@ pub async fn run_trace_op_buffer(
         // select on either buffering an op, ticking the flush interval, or shutting down
         select! {
             Some(buffered_write) = buffer_rx.recv() => {
-                match buffered_write.request {
-                    ExportRequest::GrpcExportTraceServiceRequest(r) => {
-                        let metadata = r.2.metadata().clone();
-                        let default_session_id = tonic::metadata::MetadataValue::try_from(ider::uuid()).unwrap();
-                        let session_id = metadata
-                            .get("session_id")
-                            .unwrap_or(&default_session_id)
-                            .to_str()
-                            .unwrap();
-                        let _ = ops.insert(session_id.to_string(), ExportRequest::GrpcExportTraceServiceRequest(r));
-                        notifies.push((session_id.to_string(), buffered_write.response_tx));
-                    }
-
-                    ExportRequest::HttpJsonExportTraceServiceRequest(r) => {
-                        let session_id = ider::uuid();
-                        let _ = ops.insert(session_id.to_string(), ExportRequest::HttpJsonExportTraceServiceRequest(r));
-                        notifies.push((session_id.to_string(), buffered_write.response_tx));
-                    }
-                }
+                let session_id = ider::uuid();
+                let _ = ops.insert(session_id.clone(), buffered_write.request);
+                        notifies.push((session_id, buffered_write.response_tx));
             },
             _ = interval.tick() => {
                 if ops.is_empty() {
                     continue;
                 }
                 // send ops into IO flush channel and wait for response
-                if let Err(e) = io_flush_tx.send(ops) {
+                if let Err(e) = io_flush_tx.send(ops.clone()) {
                     log::error!("io_flush_tx send e : {}, len: {}", e, io_flush_tx.len());
                 }
 
                 match io_flush_notify_rx.recv().expect("wal io thread is dead") {
                     Ok(mut resp) => {
-                        for (sid, notify) in &resp {
-                            if let Ok((_, w, data_buf, service_name)) = notify {
-                                // todo move memtable logic from wal writer , memtable refresh
-                                let _ = crate::service::ingestion::write_memtable(w, sid, data_buf.clone(), service_name).await;
-                            }
+                        for (sid, export_request) in &ops {
+                            let ExportRequest::TraceEntry(er) = export_request;
+                            let (org_id, thread_id, stream_name, entry, _, _,_, _, _) = er;
+                            let _ = crate::service::ingestion::write_memtable(org_id, *thread_id, sid, entry.clone(), stream_name).await;
                         }
 
                         // notify the watchers of the write response
                         for (sid, response_tx) in notifies {
                             match resp.remove(&sid) {
                                 Some(r) => {
-                                    let bwr = match r {
+                                    let bwr = match r.response {
                                         Ok(ets) => {
-                                            BufferedWriteResult::Success(ets.0)
+                                            BufferedWriteResult::Success(ets)
                                         }
                                         Err(e) => {
                                             log::error!("io_flush_notify_rx resp error : {e}");
